+++ tests/functional/test_basic.py
@@ -533,49 +533,6 @@ def test_argument(pyi_builder):
     )
 
 
-@importorskip('win32com')
-def test_pywin32_win32com(pyi_builder):
-    pyi_builder.test_source(
-        """
-        # Test importing some modules from pywin32 package.
-        # All modules from pywin32 depens on module pywintypes.
-        # This module should be also included.
-        import win32com
-        import win32com.client
-        import win32com.server
-        """
-    )
-
-
-#@pytest.mark.xfail(reason="Requires post-create-package hooks (issue #1322)")
-@importorskip('win32com')
-def test_pywin32_comext(pyi_builder):
-    pyi_builder.test_source(
-        """
-        # Test importing modules from win32com that are actually present in
-        # win32comext, and made available by __path__ changes in win32com.
-        from win32com.shell import shell
-        from win32com.propsys import propsys
-        from win32com.bits import bits
-        """
-    )
-
-
-@importorskip('win32ui')
-@xfail(reason="https://github.com/mhammond/pywin32/issues/1614")
-def test_pywin32_win32ui(pyi_builder):
-    pyi_builder.test_source(
-        """
-        # Test importing some modules from pywin32 package.
-        # All modules from pywin32 depens on module pywintypes.
-        # This module should be also included.
-        import win32ui
-        from pywin.mfc.dialog import Dialog
-        d = Dialog(win32ui.IDD_SIMPLE_INPUT)
-        """
-    )
-
-
 @pytest.mark.win32
 def test_renamed_exe(pyi_builder):
     _old_find_executables = pyi_builder._find_executables
